backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==3.2.2",
    "fastapi[all]>=0.136.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "sqlalchemy[asyncio]>=2.0.49",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
